Tax calculation added

+++ Models/TblSetting.cs
@@ -9,5 +9,5 @@ public partial class TblSetting
 
     public string? Property { get; set; }
 
-    public double? Value { get; set; }
+    public decimal? Value { get; set; }
 }
